# public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "KanBlam API",
    "version": "1.0",
    "description": "Token-authenticated REST API over your KanBlam workspace. Create a personal access token in Settings → API tokens and send it as `Authorization: Bearer kb_…`. A token acts as its user, limited by its scopes (`read`, `write`). Rate limit: 120 requests/minute per token."
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerToken": {
        "type": "http",
        "scheme": "bearer",
        "description": "Personal access token from Settings → API tokens (kb_…)."
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string",
                "enum": [
                  "unauthorized",
                  "forbidden",
                  "not_found",
                  "invalid_request",
                  "rate_limited",
                  "internal"
                ]
              },
              "message": {
                "type": "string"
              }
            },
            "required": [
              "code",
              "message"
            ]
          }
        },
        "required": [
          "error"
        ]
      }
    }
  },
  "paths": {
...
    "/tasks/{id}": {
      "get": {
        "tags": [
          "Tasks"
        ],
        "summary": "Get a task",
        "x-required-scope": "read",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Task id"
          }
        ],
        "responses": {
          "200": {
            "description": "Task",
            "content": {
              "application/json": {
                "example": {
                  "task": {
                    "id": "cku…task",
                    "name": "Tune 2 m dipole antenna",
                    "description": null,
                    "notes": null,
                    "project": {
                      "id": "cku…proj",
                      "name": "Payload & Electronics",
                      "code": "PAY"
                    },
                    "stage": {
                      "id": "cku…stage",
                      "name": "In Progress",
                      "isTerminal": false
                    },
                    "priority": {
                      "id": "cku…prio",
                      "name": "Medium"
                    },
                    "assignee": {
                      "id": "cku…user",
                      "name": "Brisk Otter"
                    },
                    "isImportant": true,
                    "isUrgent": false,
                    "quadrant": "q2",
                    "startDate": null,
                    "dueDate": "2026-08-01",
                    "progressPct": 50,
                    "progressManual": false,
                    "recurring": false,
                    "tags": [
                      {
                        "id": "cku…tag",
                        "name": "ham-radio",
                        "color": "#8b5cf6"
                      }
                    ],
                    "subtasks": [
                      {
                        "id": "cku…sub",
                        "title": "Trim both legs",
                        "completed": true,
                        "position": 0
                      }
                    ],
                    "createdAt": "2026-07-10T18:37:13.553Z",
                    "updatedAt": "2026-07-10T18:37:13.553Z"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Missing, malformed, revoked, or expired token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Token lacks the required scope",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Not found (cross-workspace ids are indistinguishable from unknown ids)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded — see Retry-After",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerToken": []
          }
        ]
      },
      "patch": {
        "tags": [
          "Tasks"
        ],
        "summary": "Update a task",
        "description": "Partial update. For recurring instances this edits THIS occurrence only; series editing is not part of API v1. `tagIds` replaces the full tag set.",
        "x-required-scope": "write",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Task id"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "projectId": {
                    "type": "string",
                    "minLength": 1
                  },
                  "name": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 500
                  },
                  "description": {
                    "anyOf": [
                      {
                        "type": "string",
                        "maxLength": 10000
                      },
                      {
                        "type": "null"
                      }
                    ]
                  },
                  "isImportant": {
                    "type": "boolean"
                  },
                  "isUrgent": {
                    "type": "boolean"
                  },
                  "priorityId": {
                    "type": "string",
                    "minLength": 1
                  },
                  "kanbanStageId": {
                    "type": "string",
                    "minLength": 1
                  },
                  "assigneeId": {
                    "anyOf": [
                      {
                        "type": "string",
                        "minLength": 1
                      },
                      {
                        "type": "null"
                      }
                    ]
                  },
                  "tagIds": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  },
                  "startDate": {
                    "anyOf": [
                      {
                        "type": "string"
                      },
                      {
                        "type": "null"
                      }
                    ]
                  },
                  "dueDate": {
                    "anyOf": [
                      {
                        "type": "string"
                      },
                      {
                        "type": "null"
                      }
                    ]
                  },
                  "progressPct": {
                    "type": "integer",
                    "minimum": 0,
                    "maximum": 100
                  },
                  "notes": {
                    "anyOf": [
                      {
                        "type": "string",
                        "maxLength": 50000
                      },
                      {
                        "type": "null"
                      }
                    ]
                  },
                  "progressManual": {
                    "type": "boolean"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated",
            "content": {
              "application/json": {
                "example": {
                  "task": {
                    "id": "cku…task",
                    "name": "Tune 2 m dipole antenna",
                    "description": null,
                    "notes": null,
                    "project": {
                      "id": "cku…proj",
                      "name": "Payload & Electronics",
                      "code": "PAY"
                    },
                    "stage": {
                      "id": "cku…stage",
                      "name": "In Progress",
                      "isTerminal": false
                    },
                    "priority": {
                      "id": "cku…prio",
                      "name": "Medium"
                    },
                    "assignee": {
                      "id": "cku…user",
                      "name": "Brisk Otter"
                    },
                    "isImportant": true,
                    "isUrgent": false,
                    "quadrant": "q2",
                    "startDate": null,
                    "dueDate": "2026-08-01",
                    "progressPct": 50,
                    "progressManual": false,
                    "recurring": false,
                    "tags": [
                      {
                        "id": "cku…tag",
                        "name": "ham-radio",
                        "color": "#8b5cf6"
                      }
                    ],
                    "subtasks": [
                      {
                        "id": "cku…sub",
                        "title": "Trim both legs",
                        "completed": true,
                        "position": 0
                      }
                    ],
                    "createdAt": "2026-07-10T18:37:13.553Z",
                    "updatedAt": "2026-07-10T18:37:13.553Z"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Missing, malformed, revoked, or expired token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Token lacks the required scope",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Not found (cross-workspace ids are indistinguishable from unknown ids)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "Request failed validation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded — see Retry-After",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerToken": []
          }
        ]
      },
      "delete": {
        "tags": [
          "Tasks"
        ],
        "summary": "Delete a task",
        "x-required-scope": "write",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Task id"
          }
        ],
        "responses": {
          "200": {
            "description": "Deleted",
            "content": {
              "application/json": {
                "example": {
                  "ok": true
                }
              }
            }
          },
          "401": {
            "description": "Missing, malformed, revoked, or expired token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Token lacks the required scope",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Not found (cross-workspace ids are indistinguishable from unknown ids)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded — see Retry-After",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerToken": []
          }
        ]
      }
    },
...
  }
}